refactor: integrate strategy pattern with dependency injection for Plaid
- Renamed Plaid service to implement `IBankService` interface
- Created `PlaidBankService` as a concrete strategy
- Injected `IBankService` into tRPC context in `server/context.ts`
- Refactored `userRouter`, `txRouter`, and `bankRouter` to use `ctx.bankService`
- Updated frontend usages of Plaid endpoints to use the new `bank` router namespace
