+++ package.json
@@ -33,7 +33,6 @@
     "papaparse": "^5.5.3",
     "plaid": "40.0.0",
     "postcss": "^8.5.6",
-    "prisma-json-types-generator": "^4.0.0",
     "react": "19.2.1",
     "react-dom": "19.2.1",
     "react-plaid-link": "^4.1.1",
@@ -63,6 +62,7 @@
     "prettier": "^3.7.4",
     "prettier-plugin-tailwindcss": "^0.7.2",
     "prisma": "^6.15.0",
+    "prisma-json-types-generator": "^4.0.0",
     "tailwindcss": "^4.1.17",
     "tsx": "^4.21.0",
     "typescript": "5.9.3",
