+++ pages/api/trpc/[trpc].ts
@@ -1,3 +1,4 @@
+import { createContext } from "../../../server/context";
 import { appRouter } from "../../../server/routers/_app";
 
 import * as trpcNext from "@trpc/server/adapters/next";
@@ -6,5 +7,5 @@ import * as trpcNext from "@trpc/server/adapters/next";
 // @see https://trpc.io/docs/api-handler
 export default trpcNext.createNextApiHandler({
   router: appRouter,
-  createContext: () => ({}),
+  createContext,
 });
