--- server/routers/plaid.ts
@@ -1,12 +1,0 @@
-import { createPlaidLinkToken } from "../services/plaid";
-
-import { procedure, router } from "server/trpc";
-import { z } from "zod";
-
-const plaidRouter = router({
-  createLinkToken: procedure.input(z.void()).query(async () => {
-    return await createPlaidLinkToken();
-  }),
-});
-
-export default plaidRouter;
